--- src/firebase/index.js
@@ -1,5 +1,0 @@
-import { writable } from "svelte/store";
-
-export let analytics = writable(0);
-
-export let performance = writable(0);
