--- src/graphql/store.js
@@ -1,5 +1,0 @@
-import { writable } from "svelte/store";
-
-export const serverUri = writable("");
-
-export const queries = writable({});
