+++ src/utils.js
@@ -2,10 +2,7 @@ import { onDestroy } from "svelte";
 
 export function onInterval(callback, milliseconds, condition) {
   const interval = setInterval(() => {
-    let cond = condition;
-    if (typeof condition.subscribe === "function")
-      condition.subscribe((c) => (cond = c))();
-    if (!cond) return clearInterval(interval);
+    if (!readStore(condition)) return clearInterval(interval);
     return callback();
   }, milliseconds);
   onDestroy(() => {
